chore: address PR review feedback on projects
- mint a new chat's id in a `useState` initializer and key the hydrate
  handler on the chat's identity, so React dropping a memo cache can't
  remount a chat the user is mid-way through composing
- drop the mention of documents from the Move to project copy now that
  project knowledge is gone
- extract the emoji picker's state into `useEmojiPickerState` so the body
  is display-only and the load/scroll logic is testable without a popover
- remove a describe block duplicated during an earlier merge
