+++ src/chats/detail.tsx
@@ -4,7 +4,7 @@
 
 import ChatUI from '@/components/chat/chat-ui'
 import { useHydrateChatStore } from './use-hydrate-chat-store'
-import { type PropsWithChildren, useEffect, useMemo } from 'react'
+import { type PropsWithChildren, useEffect, useState } from 'react'
 import { SavePartialAssistantMessagesHandler } from './save-partial-assistant-messages-handler'
 import { useParams, useSearchParams } from 'react-router'
 import { v7 as uuidv7 } from 'uuid'
@@ -13,13 +13,21 @@ import { loadChatMessageList } from '@/components/chat/chat-messages-loader'
 import { PageFallback } from '@/loading'
 
 type ChatHydrateHandlerProps = PropsWithChildren<{
-  id: string
-  isNew: boolean
+  /** The thread from the route, or null for a chat that doesn't exist yet. */
+  existingId: string | null
   /** Project this chat starts in, from `?projectId=`. New chats only. */
   projectId: string | null
 }>
 
-const ChatHydrateHandler = ({ children, id, isNew, projectId }: ChatHydrateHandlerProps) => {
+const ChatHydrateHandler = ({ children, existingId, projectId }: ChatHydrateHandlerProps) => {
+  const isNew = existingId === null
+  // A new chat's id is minted here, once per mount: `useState`'s initializer is a
+  // guarantee, where a `useMemo` is only a hint. React is free to drop a memo
+  // cache and recompute, which would mint a second id for the same chat and
+  // remount the thread the user is mid-way through composing. The mount itself is
+  // keyed by the route below, so a genuinely different chat gets a fresh id.
+  const [id] = useState(() => existingId ?? uuidv7())
+
   const { hydrateChatStore, isReady, saveMessages, saveStreamingMessage } = useHydrateChatStore({
     id,
     isNew,
@@ -53,16 +61,15 @@ export default function ChatDetailPage() {
   const params = useParams()
   const [searchParams] = useSearchParams()
 
-  const isNew = params.chatThreadId === 'new'
+  const chatThreadId = params.chatThreadId
+  const isNew = chatThreadId === 'new'
 
   // A new chat's project comes from the URL, and nothing is persisted yet, so the
-  // query string is part of the chat's identity. Without it in the deps, going
-  // from `/chats/new?projectId=X` to `/chats/new` reused the same id: no remount,
-  // no re-hydration, and the session kept project X — silently filing the first
-  // message under a project the user had just left.
+  // query string is part of the chat's identity. Without it in the key, going
+  // from `/chats/new?projectId=X` to `/chats/new` reused the same chat: no
+  // remount, no re-hydration, and the session kept project X — silently filing
+  // the first message under a project the user had just left.
   const projectId = isNew ? searchParams.get('projectId') : null
-
-  const id = useMemo(() => (isNew ? uuidv7() : params.chatThreadId || null), [isNew, params.chatThreadId, projectId])
 
   // Warm the lazily-split message subtree for new chats as well as existing
   // ones: the first send swaps the empty state for the message list mid-
@@ -73,12 +80,19 @@ export default function ChatDetailPage() {
     void loadChatMessageList()
   }, [])
 
-  if (!id) {
+  if (!chatThreadId) {
     return null
   }
 
   return (
-    <ChatHydrateHandler key={id} id={id} isNew={isNew} projectId={projectId}>
+    // Keyed on the chat's identity rather than its id: the route's thread, or "a
+    // new chat in this project". Navigating between two of those remounts the
+    // handler, which is what re-hydrates the store and re-mints a new chat's id.
+    <ChatHydrateHandler
+      key={isNew ? `new:${projectId ?? ''}` : chatThreadId}
+      existingId={isNew ? null : chatThreadId}
+      projectId={projectId}
+    >
       <ChatUI />
     </ChatHydrateHandler>
   )
