+++ src/projects/emoji-picker.tsx
@@ -23,8 +23,8 @@
  */
 
 import { Search, Smile, X } from 'lucide-react'
-import { useEffect, useMemo, useRef, useState } from 'react'
-import { Virtualizer, type VirtualizerHandle } from 'virtua'
+import { useState } from 'react'
+import { Virtualizer } from 'virtua'
 
 import { Button } from '@/components/ui/button'
 import { Input } from '@/components/ui/input'
@@ -32,21 +32,11 @@ import { MobileActionSheet } from '@/components/ui/mobile-action-sheet'
 import { Popover, PopoverContent, PopoverTrigger } from '@/components/ui/popover'
 import { useIsMobile } from '@/hooks/use-mobile'
 import { cn } from '@/lib/utils'
-import {
-  filterCatalog,
-  loadEmojiCatalog,
-  suggestedEmoji,
-  toRows,
-  type EmojiCategory,
-  type EmojiEntry,
-} from './emoji-catalog'
+import { suggestedEmoji } from './emoji-catalog'
+import { useEmojiPickerState } from './use-emoji-picker-state'
 
 /** Height reserved for the pinned category label, in pixels. */
 const pinnedLabelHeight = 24
-
-/** Emoji per row. Desktop fits 9 in a 19rem popover; the sheet is full-width. */
-const desktopPerRow = 9
-const mobilePerRow = 8
 
 type EmojiPickerProps = {
   /** Current emoji, or null for the default folder glyph. */
@@ -87,85 +77,27 @@ const EmojiButton = ({
 )
 
 /**
- * Flatten categories into emoji rows, plus the category each row belongs to.
- *
- * The label is carried *alongside* the rows rather than as a heading row in the
- * list, because the heading has to stay pinned while its category scrolls (see
- * `EmojiPickerBody`). A heading rendered in flow can't do that here: `virtua`
- * unmounts rows that leave the viewport, so `position: sticky` on one would
- * vanish the moment it scrolled out — which is exactly when it needs to stick.
- */
-const toGridRows = (
-  categories: readonly EmojiCategory[],
-  perRow: number,
-): { rows: EmojiEntry[][]; labelByRow: string[] } => {
-  const rows: EmojiEntry[][] = []
-  const labelByRow: string[] = []
-  for (const category of categories) {
-    for (const entries of toRows(category.emoji, perRow)) {
-      rows.push(entries)
-      labelByRow.push(category.label)
-    }
-  }
-  return { rows, labelByRow }
-}
-
-/**
  * Search field + suggested row + virtualized full set. Shared by both shells so
- * the two presentations can't drift in behaviour.
+ * the two presentations can't drift in behaviour. All state lives in
+ * `useEmojiPickerState`, leaving this component to presentation only.
  */
 const EmojiPickerBody = ({
   value,
   onChange,
   isMobile,
 }: Pick<EmojiPickerProps, 'value' | 'onChange'> & { isMobile: boolean }) => {
-  const [search, setSearch] = useState('')
-  // `'failed'` rather than a third `useState`: the load has three outcomes, and one
-  // slot keeps them mutually exclusive.
-  const [catalog, setCatalog] = useState<EmojiCategory[] | 'failed' | null>(null)
-  const scrollRef = useRef<HTMLDivElement>(null)
-  const virtualizerRef = useRef<VirtualizerHandle>(null)
-  const perRow = isMobile ? mobilePerRow : desktopPerRow
-
-  // Legitimate effect: fetching from an external module on mount. The catalogue
-  // module memoizes a *successful* load, so reopening never re-imports — and it
-  // does not cache a failure, so reopening after one really does retry.
-  useEffect(() => {
-    let cancelled = false
-    const load = async () => {
-      try {
-        const loaded = await loadEmojiCatalog()
-        if (!cancelled) {
-          setCatalog(loaded)
-        }
-      } catch (error) {
-        // A rejected dynamic import (offline, or a chunk 404 after a redeploy)
-        // would otherwise leave the body on "Loading emoji…" forever.
-        console.warn('Failed to load the emoji catalogue', error)
-        if (!cancelled) {
-          setCatalog('failed')
-        }
-      }
-    }
-    load()
-    return () => {
-      cancelled = true
-    }
-  }, [])
-
-  const { rows, labelByRow } = useMemo(
-    () =>
-      catalog && catalog !== 'failed'
-        ? toGridRows(filterCatalog(catalog, search), perRow)
-        : { rows: [], labelByRow: [] },
-    [catalog, search, perRow],
-  )
-  const showSuggested = search.trim().length === 0
-  // The category label pinned at the top of the grid. Derived from the scroll
-  // offset via the virtualizer rather than tracked per row, so it stays correct
-  // even when the rows it describes have been unmounted.
-  const [topRow, setTopRow] = useState(0)
-  const pinnedLabel = showSuggested ? labelByRow[Math.min(topRow, labelByRow.length - 1)] : undefined
+  const {
+    search,
+    setSearch,
+    catalog,
+    rows,
+    perRow,
+    showSuggested,
+    pinnedLabel,
+    scrollRef,
+    virtualizerRef,
+    handleScroll,
+  } = useEmojiPickerState(isMobile)
   const gridClass = isMobile ? 'grid grid-cols-8 gap-1' : 'grid grid-cols-9 gap-1'
 
   return (
@@ -231,7 +163,7 @@ const EmojiPickerBody = ({
             <Virtualizer
               ref={virtualizerRef}
               scrollRef={scrollRef}
-              onScroll={(offset) => setTopRow(virtualizerRef.current?.findItemIndex(offset) ?? 0)}
+              onScroll={handleScroll}
               // Clears the first row of the pinned label instead of starting
               // underneath it.
               startMargin={pinnedLabel ? pinnedLabelHeight : 0}
